requirements.txt
# Web framework
fastapi==0.115.0
uvicorn==0.30.6

# Scheduler
apscheduler==3.10.4

# Database
asyncpg==0.29.0

# AI
anthropic==0.34.0
groq==0.11.0

# Market data
yfinance==0.2.43
FinanceDataReader==0.9.50
pykrx==1.0.47

# Macro data
fredapi==0.5.2

# HTTP / scraping
httpx==0.27.2
beautifulsoup4==4.12.3
feedparser==6.0.11

# Utils
python-dotenv==1.0.1
pydantic==2.9.2
